requirements.txt
asgiref==3.5.2
autopep8==1.7.0
backports.zoneinfo==0.2.1;python_version<"3.9"
certifi==2022.9.24
charset-normalizer==2.1.1
cloudinary==1.30.0
dj-database-url==1.0.0
Django==4.1.1
django-ckeditor==6.5.1
django-cloudinary-storage==0.3.0
django-heroku==0.3.1
django-js-asset==2.0.0
gunicorn==20.1.0
idna==3.4
Pillow==9.2.0
psycopg2==2.9.3
pycodestyle==2.9.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
tzdata==2022.2
urllib3==1.26.12
whitenoise==6.2.0
